+++ openAIClient/openAIClient/mainWindow.py
@@ -1,9 +1,8 @@
 import sys
 import os
 import logging
-import configparser
 import datetime
-from typing import Optional, Tuple
+from typing import Optional
 from pathlib import Path
 
 from PySide6.QtWidgets import (
@@ -14,7 +13,7 @@ from PySide6.QtCore import QByteArray, QSettings, Qt, QPoint
 from PySide6.QtGui import QIcon, QKeySequence, QAction
 
 from openAIClient.gui.Ui_mainWindow import Ui_MainWindow
-from openAIClient.gui.chatTab import ChatTab
+from openAIClient.chatTab import ChatTab
 from openAIClient.chatSession import ChatSession
 from openAIClient.utils.database import ConversationDatabase
 from openAIClient.config import getDatabasePath, resourcePath, getDeploymentConfigs
